extract jumps
